Move up level sorting and limitation of locations
Moved from query level to datasource level.
Reason: fix bug with type-specific select.

+++ shared/src/commonMain/kotlin/ru/z8/louttsev/cheaptripmobile/shared/infrastructure/persistence/LocationDb.kt
@@ -27,8 +27,8 @@ class LocationDb(sqlDriver: SqlDriver) : LocalDbStorage<Location>(sqlDriver) {
 
     override fun get(parameters: ParamsBundle): List<Location> {
         val needle = parameters.get(Key.NEEDLE) as String
-        val limit = parameters.get(Key.LIMIT) as Long
+        val limit = parameters.get(Key.LIMIT) as Int
 
-        return selectLocationsByName(needle, limit)
+        return selectLocationsByName(needle, limit.toLong())
     }
 }
