+++ shared/src/commonMain/kotlin/ru/z8/louttsev/cheaptripmobile/shared/model/LocationRepository.kt
@@ -38,7 +38,7 @@ class LocationRepository(
     fun searchLocationsByName(
         needle: String,
         type: Type = ALL,
-        limit: Long = 10,
+        limit: Int = 10,
         locale: Locale = currentLocale
     ): List<Location> {
         val params = ParamsBundle().apply {
